chore: update version of @salesforce/core (#129)
@W-8853146@
