+++ lerna.json
@@ -3,5 +3,5 @@
   "packages": [
     "packages/*"
   ],
-  "version": "0.1.12"
+  "version": "0.1.13"
 }
