+++ packages/apex-node/package.json
@@ -1,18 +1,19 @@
 {
   "name": "@salesforce/apex-node",
   "description": "Salesforce js library for Apex",
-  "version": "0.1.12",
+  "version": "0.1.13",
   "author": "Salesforce",
   "bugs": "https://github.com/forcedotcom/salesforcedx-apex/issues",
   "main": "lib/src/index.js",
   "dependencies": {
-    "@salesforce/core": "2.11.0",
+    "@salesforce/core": "2.12.0",
     "faye": "1.1.3"
   },
   "devDependencies": {
     "@salesforce/ts-sinon": "^1.1.2",
     "@salesforce/ts-types": "1.2.2",
     "@types/chai": "^4",
+    "@types/mkdirp": "0.5.2",
     "@types/mocha": "^5",
     "@types/node": "^10",
     "@types/sinon": "^7.5.2",
