+++ packages/apex-node/test/execute/executeService.test.ts
@@ -32,11 +32,17 @@ describe('Apex Execute Tests', async () => {
     $$.setConfigStubContents('AuthInfoConfig', {
       contents: await testData.getConfig()
     });
+    // Stub retrieveMaxApiVersion to get over "Domain Not Found: The org cannot be found" error
+    sandboxStub
+      .stub(Connection.prototype, 'retrieveMaxApiVersion')
+      .resolves('50.0');
+
     mockConnection = await Connection.create({
       authInfo: await AuthInfo.create({
         username: testData.username
       })
     });
+
     sandboxStub.stub(fs, 'readFileSync').returns('System.assert(true);');
     fsStub = sandboxStub.stub(fs, 'existsSync').returns(true);
   });
