+++ packages/apex-node/test/streaming/streamingClient.test.ts
@@ -35,6 +35,10 @@ describe('Streaming API Client', () => {
     $$.setConfigStubContents('AuthInfoConfig', {
       contents: await testData.getConfig()
     });
+    // Stub retrieveMaxApiVersion to get over "Domain Not Found: The org cannot be found" error
+    sandboxStub
+      .stub(Connection.prototype, 'retrieveMaxApiVersion')
+      .resolves('50.0');
     mockConnection = await Connection.create({
       authInfo: await AuthInfo.create({
         username: testData.username
