+++ packages/plugin-apex/test/commands/force/apex/test/report.test.ts
@@ -10,7 +10,7 @@ import { Messages, SfdxProject } from '@salesforce/core';
 import * as fs from 'fs';
 import * as path from 'path';
 import * as stream from 'stream';
-import { createSandbox, SinonSandbox } from 'sinon';
+import { createSandbox, SinonSandbox, SinonStub } from 'sinon';
 import {
   testRunSimple,
   cliJsonResult,
@@ -372,7 +372,7 @@ describe('force:apex:test:report', () => {
     .it(
       'should create test-run-codecoverage file with correct content when code cov is specified',
       ctx => {
-        expect(ctx.myStub.args).to.deep.equal([
+        expect((ctx.myStub as SinonStub).args).to.deep.equal([
           [
             runWithCoverage,
             {
